# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "D&D API",
    "version": "1.0.0",
    "description": "Documentación de la API para gestión de campañas y fichas de D&D"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Servidor Local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Introduce el token JWT recibido en el login (sin el prefijo \"Bearer \")"
      }
    },
    "schemas": {
...
      "Language": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID de MongoDB del idioma."
          },
          "name": {
            "type": "string",
            "description": "Nombre del idioma."
          },
          "type": {
            "type": "string",
            "description": "Tipo de idioma (ej. Estándar, Exótico)."
          },
          "description": {
            "type": "string",
            "description": "Descripción del idioma."
          },
          "script": {
            "type": "string",
            "description": "Alfabeto o escritura del idioma."
          },
          "ruleset": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Array de IDs de los sistemas a los que pertenece el idioma."
          }
        }
      },
      "InputCreateLanguage": {
        "type": "object",
        "required": [
          "name",
          "ruleset"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "Nombre del idioma."
          },
          "description": {
            "type": "string",
            "description": "Descripción."
          },
          "type": {
            "type": "string",
            "description": "Tipo de idioma."
          },
          "script": {
            "type": "string",
            "description": "Alfabeto del idioma."
          },
          "ruleset": {
            "type": "string",
            "description": "ID o nombre del sistema inicial asociado al idioma."
          }
        }
      },
      "InputUpdateLanguage": {
        "type": "object",
        "required": [
          "id"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "ID del idioma a actualizar."
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "script": {
            "type": "string"
          }
        }
      },
      "Attribute": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID de MongoDB de la característica."
          },
          "ruleset": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Array de IDs o nombres de sistemas asociados."
          },
          "name": {
            "type": "string",
            "description": "Nombre de la característica (ej. Fuerza)."
          },
          "description": {
            "type": "string",
            "description": "Descripción detallada."
          },
          "key": {
            "type": "string",
            "description": "Clave identificadora única (ej. str, dex)."
          },
          "abbreviation": {
            "type": "string",
            "description": "Abreviatura (ej. FUE, DES)."
          },
          "icon": {
            "type": "string",
            "description": "Nombre o URL del icono opcional."
          }
        }
      },
      "InputCreateAttribute": {
        "type": "object",
        "required": [
          "ruleset",
          "name",
          "description",
          "key",
          "abbreviation"
        ],
        "properties": {
          "ruleset": {
            "type": "string",
            "description": "ID o nombre del sistema inicial desde el que se crea."
          },
          "name": {
            "type": "string",
            "description": "Nombre de la característica."
          },
          "description": {
            "type": "string",
            "description": "Descripción."
          },
          "key": {
            "type": "string",
            "description": "Clave identificadora única."
          },
          "abbreviation": {
            "type": "string",
            "description": "Abreviatura."
          },
          "icon": {
            "type": "string",
            "description": "Nombre o URL del icono opcional."
          }
        }
      },
      "InputUpdateAttribute": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "key": {
            "type": "string"
          },
          "abbreviation": {
            "type": "string"
          },
          "icon": {
            "type": "string"
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/languages": {
      "get": {
        "summary": "Obtener los idiomas asociados a ciertos sistemas",
        "tags": [
          "Idiomas"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "ruleset",
            "required": false,
            "schema": {
              "oneOf": [
                {
                  "type": "string"
                },
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              ]
            },
            "description": "Sistema o sistemas para los cuales filtrar los idiomas."
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de idiomas obtenida con éxito.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Language"
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autorizado."
          },
          "500": {
            "description": "Error interno del servidor."
          }
        }
      },
      "post": {
        "summary": "Crear un nuevo idioma",
        "tags": [
          "Idiomas"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InputCreateLanguage"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Idioma creado con éxito.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Language"
                }
              }
            }
          },
          "400": {
            "description": "Parámetros incorrectos o idioma ya existente."
          },
          "401": {
            "description": "No autorizado."
          },
          "409": {
            "description": "Conflicto, el idioma ya existe."
          },
          "500": {
            "description": "Error interno del servidor."
          }
        }
      }
    },
    "/languages/{id}": {
      "put": {
        "summary": "Actualizar un idioma existente",
        "tags": [
          "Idiomas"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID del idioma a actualizar."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InputUpdateLanguage"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Idioma actualizado con éxito.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Language"
                }
              }
            }
          },
          "400": {
            "description": "Parámetros incorrectos."
          },
          "401": {
            "description": "No autorizado."
          },
          "404": {
            "description": "Idioma no encontrado."
          },
          "500": {
            "description": "Error interno del servidor."
          }
        }
      },
      "delete": {
        "summary": "Realizar un borrado lógico de un idioma",
        "tags": [
          "Idiomas"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID del idioma a borrar."
          }
        ],
        "responses": {
          "204": {
            "description": "Idioma borrado exitosamente."
          },
          "400": {
            "description": "ID de idioma requerido."
          },
          "403": {
            "description": "No tienes permisos para borrar este idioma."
          },
          "404": {
            "description": "Idioma no encontrado."
          },
          "500": {
            "description": "Error del servidor."
          }
        }
      }
    },
    "/languages/{id}/restore": {
      "patch": {
        "summary": "Restaurar un idioma borrado lógicamente",
        "tags": [
          "Idiomas"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID del idioma a restaurar."
          }
        ],
        "responses": {
          "200": {
            "description": "Idioma restaurado exitosamente."
          },
          "400": {
            "description": "ID de idioma requerido."
          },
          "403": {
            "description": "No tienes permisos para restaurar este idioma."
          },
          "404": {
            "description": "Idioma no encontrado."
          },
          "500": {
            "description": "Error del servidor."
          }
        }
      }
    },
    "/attributes": {
      "post": {
        "summary": "Crear una nueva característica",
        "tags": [
          "Caracteristicas"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InputCreateAttribute"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Característica creada con éxito.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Attribute"
                }
              }
            }
          },
          "400": {
            "description": "Faltan campos obligatorios."
          },
          "401": {
            "description": "No autorizado."
          },
          "409": {
            "description": "Ya existe una característica con esta clave en el sistema."
          },
          "500": {
            "description": "Error del servidor."
          }
        }
      }
    },
    "/attributes/{id}": {
      "put": {
        "summary": "Actualizar una característica existente",
        "tags": [
          "Caracteristicas"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID de la característica a editar."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InputUpdateAttribute"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Característica modificada con éxito.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Attribute"
                }
              }
            }
          },
          "400": {
            "description": "Falta el ID de la característica."
          },
          "401": {
            "description": "No autorizado."
          },
          "404": {
            "description": "No se encontró la característica."
          },
          "500": {
            "description": "Error del servidor."
          }
        }
      },
      "delete": {
        "summary": "Realizar un borrado lógico de una característica",
        "tags": [
          "Caracteristicas"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID de la característica a borrar."
          }
        ],
        "responses": {
          "204": {
            "description": "Característica borrada exitosamente."
          },
          "400": {
            "description": "ID de característica requerido."
          },
          "403": {
            "description": "No tienes permisos para borrar esta característica."
          },
          "404": {
            "description": "Característica no encontrada."
          },
          "500": {
            "description": "Error del servidor."
          }
        }
      }
    },
    "/attributes/{id}/restore": {
      "patch": {
        "summary": "Restaurar una característica borrada lógicamente",
        "tags": [
          "Caracteristicas"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID de la característica a restaurar."
          }
        ],
        "responses": {
          "200": {
            "description": "Característica restaurada exitosamente."
          },
          "400": {
            "description": "ID de característica requerido."
          },
          "403": {
            "description": "No tienes permisos para restaurar esta característica."
          },
          "404": {
            "description": "Característica no encontrada."
          },
          "500": {
            "description": "Error del servidor."
          }
        }
      }
    }
  },
...
}